+++ e2e/nextjs-routes.d.ts
@@ -36,6 +36,7 @@ declare module "next/link" {
   export * from "next/dist/client/link";
 
   type Query = { query?: { [key: string]: string | string[] | undefined } };
+  type StaticRoute = Exclude<Route, { query: any }>["pathname"];
 
   export interface LinkProps<Href extends Route | Query = Route | Query>
     extends Omit<NextLinkProps, "href" | "locale"> {
@@ -57,7 +58,7 @@ declare module "next/link" {
     props: PropsWithChildren<LinkProps<Route>>
   ): LinkReactElement;
   declare function Link(
-    props: PropsWithChildren<LinkProps<Route["pathname"]>>
+    props: PropsWithChildren<LinkProps<StaticRoute>>
   ): LinkReactElement;
   declare function Link(
     props: PropsWithChildren<LinkProps<Query>>
@@ -74,6 +75,7 @@ declare module "next/router" {
   export { default } from "next/dist/client/router";
 
   type NextTransitionOptions = NonNullable<Parameters<Router["push"]>[2]>;
+  type StaticRoute = Exclude<Route, { query: any }>["pathname"];
 
   interface TransitionOptions extends Omit<NextTransitionOptions, "locale"> {
     locale?: false;
@@ -100,7 +102,7 @@ declare module "next/router" {
       options?: TransitionOptions
     ): Promise<boolean>;
     push(
-      url: Route["pathname"],
+      url: StaticRoute,
       as?: string,
       options?: TransitionOptions
     ): Promise<boolean>;
@@ -116,7 +118,7 @@ declare module "next/router" {
       options?: TransitionOptions
     ): Promise<boolean>;
     replace(
-      url: Route["pathname"],
+      url: StaticRoute,
       as?: string,
       options?: TransitionOptions
     ): Promise<boolean>;